06-stream/homework/requirements.txt
wget~=3.2
pyyaml~=6.0
kafka-python~=2.0.2
fastavro~=1.7.3
confluent-kafka~=2.0.2
jupyterlab~=3.5.1
pandas~=1.5.2